requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.51.0
anyio==4.9.0
argcomplete==3.6.2
asyncer==0.0.7
attrs==25.3.0
backoff==2.2.1
bidict==0.23.1
boto3==1.38.23
botocore==1.38.23
cachetools==5.5.2
certifi==2025.4.26
chainlit==2.5.5
charset-normalizer==3.4.2
chevron==0.14.0
click==8.1.8
cohere==5.15.0
colorama==0.4.6
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
eval_type_backport==0.2.2
executing==2.2.0
fasta2a==0.2.6
fastapi==0.115.12
fastavro==1.11.1
filelock==3.18.0
filetype==1.2.0
frozenlist==1.6.0
fsspec==2025.3.2
google-auth==2.40.2
google-genai==1.16.1
googleapis-common-protos==1.70.0
griffe==1.7.3
groq==0.25.0
grpcio==1.71.0
h11==0.16.0
hf-xet==1.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.1
idna==3.10
importlib_metadata==8.6.1
inflection==0.5.1
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
langfuse==2.60.5
Lazify==0.4.0
literalai==0.1.201
logfire==3.16.0
logfire-api==3.15.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.7.1
mdurl==0.1.2
mistralai==1.7.1
monotonic==1.6
multidict==6.4.3
mypy_extensions==1.1.0
nest-asyncio==1.6.0
openai==1.77.0
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-exporter-otlp-proto-http==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-alephalpha==0.40.3
opentelemetry-instrumentation-anthropic==0.40.3
opentelemetry-instrumentation-bedrock==0.40.3
opentelemetry-instrumentation-chromadb==0.40.3
opentelemetry-instrumentation-cohere==0.40.3
opentelemetry-instrumentation-crewai==0.40.3
opentelemetry-instrumentation-google-generativeai==0.40.3
opentelemetry-instrumentation-groq==0.40.3
opentelemetry-instrumentation-haystack==0.40.3
opentelemetry-instrumentation-lancedb==0.40.3
opentelemetry-instrumentation-langchain==0.40.3
opentelemetry-instrumentation-llamaindex==0.40.3
opentelemetry-instrumentation-logging==0.52b1
opentelemetry-instrumentation-marqo==0.40.3
opentelemetry-instrumentation-mcp==0.40.3
opentelemetry-instrumentation-milvus==0.40.3
opentelemetry-instrumentation-mistralai==0.40.3
opentelemetry-instrumentation-ollama==0.40.3
opentelemetry-instrumentation-openai==0.40.3
opentelemetry-instrumentation-pinecone==0.40.3
opentelemetry-instrumentation-qdrant==0.40.3
opentelemetry-instrumentation-replicate==0.40.3
opentelemetry-instrumentation-requests==0.52b1
opentelemetry-instrumentation-sagemaker==0.40.3
opentelemetry-instrumentation-sqlalchemy==0.52b1
opentelemetry-instrumentation-threading==0.52b1
opentelemetry-instrumentation-together==0.40.3
opentelemetry-instrumentation-transformers==0.40.3
opentelemetry-instrumentation-urllib3==0.52b1
opentelemetry-instrumentation-vertexai==0.40.3
opentelemetry-instrumentation-watsonx==0.40.3
opentelemetry-instrumentation-weaviate==0.40.3
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-semantic-conventions-ai==0.4.5
opentelemetry-util-http==0.52b1
packaging==24.2
pipdeptree==2.26.1
posthog==3.25.0
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic-ai==0.2.6
pydantic-ai-slim==0.2.6
pydantic-evals==0.2.6
pydantic-graph==0.2.6
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.0
python-multipart==0.0.18
python-socketio==5.13.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rsa==4.9.1
s3transfer==0.13.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sse-starlette==2.3.4
starlette==0.41.3
syncer==2.0.3
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.1
tomli==2.2.1
tqdm==4.67.1
traceloop-sdk==0.40.3
types-requests==2.32.0.20250515
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
uptrace==1.31.0
urllib3==2.4.0
uvicorn==0.34.2
watchfiles==0.20.0
wcwidth==0.2.13
websockets==15.0.1
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.0
zipp==3.21.0
